Change in downsampling and fixing raster2D and droppers
The downsampling is gridAverage instead of random.
Raster2D only worked if the cloud has angle. Now, if it does not have angle, it generate a angle with all 0.
In DroppersExtraction, sometimos when the function CheckElement(points, model) is called, points were empty, generating an error. Now it is checked.