RailsExtraction error fixed and status as .json
RailsExtraction has an error when use pdist2() to check if any mast is close to a cable. The error was that if a supposed mast had a lot of points pdist2() did not work. Now it is in a try because no mast has such a number of points.

RailwayHeuristicSegmentation_Teselas and RailwayHeuristicSegmentation know save the status in .json.  In the case of teselas also save in status a variable process. It is because sometimes there is an error that I am not able to locate.